requirements.txt
Django==2.0.2
django-environ==0.4.4
django-filter==1.1.0
django-graphql-jwt==0.1.5
flake8==3.5.0
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
iso8601==0.1.12
mccabe==0.6.1
promise==2.1
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.3
pytz==2018.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.4
